modfied docs to better support self hosting / split success page into multiple components
