+++ .env.example
@@ -9,7 +9,7 @@ GITHUB_APP_CLIENT_ID="Iv1.your-client-id-here"
 GITHUB_APP_CLIENT_SECRET="your-client-secret-here"
 NEXT_PUBLIC_GITHUB_APP_NAME="jules-task-queue"
 
-# Cron job verification token (for Firebase you also need to add thin in ./functions/.env.local)
+# Cron job verification token (for Firebase you also need to add this in ./functions/.env as well)
 CRON_SECRET="1234567890"
 
 # Development
