+++ functions/.env.example
@@ -1,5 +1,5 @@
 # Local environment variables for Firebase Functions testing
-# Copy this to .env.local and fill in your values
+# Copy this to .env and fill in your values
 
 # Secret values for local testing
 CRON_SECRET=your-local-cron-secret-for-testing
