+++ src/components/success/loading-state.tsx
@@ -1,0 +1,9 @@
+import { LoadingSpinner } from "@/components/ui/loading-spinner";
+
+export function LoadingState() {
+  return (
+    <div className="min-h-screen bg-jules-dark flex items-center justify-center px-4">
+      <LoadingSpinner size="lg" text="Processing installation..." />
+    </div>
+  );
+}
