+++ supabase/.gitignore
@@ -4,5 +4,5 @@
 
 # dotenvx
 .env.keys
-.env.local
+.env
 .env.*.local
