--- tasks.md
@@ -1,203 +1,0 @@
-# Jules Task Queueing System - Development Tasks
-
-## Phase 1: Project Foundation & Setup
-
-### Task 1: Project Initialization & Basic Structure ✅ COMPLETED
-
-- [x] Initialize Next.js 15+ project with TypeScript
-- [x] Configure Tailwind CSS and shadcn components
-- [x] Set up basic directory structure
-- [x] Create .env.example with all required variables
-- [x] Configure tsconfig.json for strict type checking
-
-### Task 2: Database Schema & Prisma Setup ✅ COMPLETED
-
-- [x] Install and configure Prisma with PostgreSQL
-- [x] Create JulesTask and WebhookLog models
-- [x] Set up database connection utilities
-- [x] Generate initial migration
-
-### Task 3: Environment & Configuration ✅ COMPLETED
-
-- [x] Create environment validation with Zod
-- [x] Set up configuration management
-- [x] Create health check endpoints
-- [x] Configure Next.js for production optimization
-
-## Phase 2: Core API Infrastructure
-
-### Task 4: tRPC Setup & Configuration ✅ COMPLETED
-
-- [x] Install and configure tRPC v10+
-- [x] Create router structure (tasks, admin, webhook)
-- [x] Set up type-safe API procedures
-- [x] Configure tRPC with Next.js App Router
-
-### Task 5: GitHub SDK Integration ✅ COMPLETED
-
-- [x] Install and configure @octokit/rest
-- [x] Create GitHub client utilities
-- [x] Implement issue and comment operations
-- [x] Add label management functions
-
-## Phase 3: Webhook System
-
-### Task 6: GitHub Webhook Handler ✅ COMPLETED
-
-- [x] Create `/api/webhooks/github/route.ts` endpoint
-- [x] Implement webhook signature verification
-- [x] Parse GitHub issue label events
-- [x] Create webhook logging system
-- [x] Add comprehensive error handling and security measures
-
-### Task 7: Task Creation & Management ✅ COMPLETED
-
-- [x] Implement `upsertJulesTask` function with enhanced schema
-- [x] Create task creation logic for 'jules' label events
-- [x] Set up 60-second delay timer for comment checking
-- [x] Add task status tracking and updates
-- [x] Implement webhook processing workflow
-
-## Phase 4: Comment Detection System
-
-### Task 8: Jules Bot Comment Analysis ✅ COMPLETED
-
-- [x] Create comment pattern detection system
-- [x] Implement Jules bot username recognition
-- [x] Add task limit detection logic
-- [x] Create success/working status detection
-- [x] Add comprehensive comment type classification
-- [x] Implement comment timestamp analysis
-- [x] Add retry logic for comment detection failures
-
-### Task 9: Workflow Decision Logic ✅ COMPLETED
-
-- [x] Implement basic comment-based decision making
-- [x] Add Human label detection and bypass
-- [x] Create workflow state management
-- [x] Add comprehensive logging
-- [x] Enhance decision tree for edge cases
-- [x] Add workflow validation and error recovery
-- [x] Implement advanced state transitions
-
-## Phase 5: Retry & Queue Management
-
-### Task 10: Retry Processing System ✅ COMPLETED
-
-- [x] Create cron job endpoint (`/api/cron/retry/route.ts`)
-- [x] Implement flagged task retry logic
-- [x] Add label swapping functionality (`jules` ↔ `jules-queue`)
-- [x] Create retry statistics and monitoring
-
-### Task 11: Cron Job & Scheduling ✅ COMPLETED
-
-- [x] Implement 30-minute retry intervals
-- [x] Add manual retry triggers for testing
-- [x] Set up Vercel cron job configuration
-- [x] Add cron job monitoring and alerting
-
-## Phase 6: tRPC API Endpoints
-
-### Task 12: Tasks Router ✅ COMPLETED
-
-- [x] Create task listing and filtering endpoints
-- [x] Add task statistics and metrics
-- [x] Implement individual task operations
-- [x] Add task status updates
-
-### Task 13: Admin Router ✅ COMPLETED
-
-- [x] Create admin-only procedures for task management
-- [x] Implement bulk retry operations
-- [x] Add system health and monitoring endpoints
-- [x] Create webhook log viewing capabilities
-
-## Phase 7: Security & Error Handling
-
-### Task 14: Security Implementation
-
-- [ ] Implement API authentication
-- [ ] Add input validation and sanitization
-- [ ] Create security headers and CORS
-
-### Task 15: Error Handling & Logging
-
-- [ ] Implement comprehensive error handling
-- [ ] Add structured logging system
-- [ ] Create error reporting and monitoring
-- [ ] Add performance tracking
-
-## Phase 8: Testing & Quality Assurance
-
-### Task 16: Unit Testing
-
-- [ ] Write tests for core functions
-- [ ] Test GitHub API integrations
-- [ ] Add database operation tests
-- [ ] Test webhook processing logic
-
-### Task 17: Integration Testing
-
-- [ ] Create end-to-end workflow tests
-- [ ] Test GitHub webhook integration
-- [ ] Add cron job testing
-- [ ] Test error scenarios and edge cases
-
-## Phase 9: Documentation & Deployment
-
-### Task 18: Documentation
-
-- [ ] Complete README with setup instructions
-- [ ] Document API endpoints and usage
-- [ ] Create deployment guide
-- [ ] Add troubleshooting documentation
-
-### Task 19: Production Deployment
-
-- [ ] Configure Vercel deployment
-- [ ] Set up production database
-- [ ] Configure GitHub webhooks
-- [ ] Set up monitoring and alerting
-
-### Task 20: Final Testing & Launch
-
-- [ ] Perform production testing
-- [ ] Validate all workflows end-to-end
-- [ ] Monitor initial production usage
-- [ ] Create launch documentation
-
-## Current Status
-
-**Phase 4 Complete** - Successfully implemented enhanced comment detection system with comprehensive pattern matching, confidence scoring, retry logic, and advanced workflow decision processing. All comment types (task_limit, working, error, completed) are now handled with robust error recovery.
-
-### Recently Completed:
-
-- ✅ **Task 8**: Enhanced Jules bot comment analysis with confidence scoring, timestamp validation, and comprehensive pattern detection for all comment types
-- ✅ **Task 9**: Advanced workflow decision logic with state validation, error recovery, and complete automated task lifecycle management
-
-### Next Up:
-
-- **Task 11**: Complete Vercel cron job configuration and monitoring
-- **Task 14-15**: Security implementation and comprehensive error handling
-
-### Key Achievements:
-
-- **Enhanced Comment Detection**: Advanced pattern matching with confidence scoring for all Jules bot response types
-- **Intelligent Workflow Processing**: Automated decision trees with error recovery and state validation
-- **Robust Retry Logic**: Exponential backoff and comprehensive error handling for comment analysis failures
-- **Complete Task Lifecycle**: Automated handling of task_limit, working, error, and completed states
-- **Enterprise Security**: Webhook signature verification and comprehensive audit logging
-- **Type-safe Infrastructure**: Full tRPC v10+ API with enhanced database schema
-- **Production-Ready Monitoring**: Comprehensive logging, metrics, and health checks
-
-## Success Criteria
-
-- [x] Webhook processes GitHub label events securely
-- [x] Tasks are created and tracked in database
-- [x] 60-second delay before checking comments
-- [x] System handles task limits and queues for retry
-- [x] Retry system processes queued tasks every 30 minutes
-- [x] All operations are logged and monitorable
-- [x] Landing page explains the system
-- [ ] Complete end-to-end workflow testing
-- [ ] Production deployment with monitoring
